learn-go-with-tests/select/race.go
package racer

import (
	"fmt"
	"net/http"
	"time"
)

/*
func Race(a, b string) string {
	aDuration := measureResponseTime(a)
	bDuration := measureResponseTime(b)

	if aDuration < bDuration {
		return a
	}
	return b
}

func measureResponseTime(url string) time.Duration {
	start := time.Now()
	http.Get(url)
	return time.Since(start)
}
*/

var tenMillionTimeout = time.Second * 10
func Race(a, b string) (string, error){
	return TimeoutRace(a, b, tenMillionTimeout)
}

func TimeoutRace(a, b string, timeout time.Duration) (string, error) {
	select {
	case <-ping(a):
		return a, nil
	case <-ping(b):
		return b, nil
	case <-time.After(timeout):
		return "", fmt.Errorf("	timeout waiting for %s and %s", a, b)
	}
}

func ping(url string) chan bool {
	ch := make(chan bool)
	go func() {
		http.Get(url)
		ch <- true
	}()
	return ch
}
